Add a bunch of debugging sys.puts.
There are issues with this when returning multiple tiddlers
in a bag, but need to delay that until the test suite is
passing. The test suite does not show the bug because it
doesn't do the thing that tickles the bug. The test suite will be
improved eventually.

Checking in now to get to stable and known state.